requirements.txt
requests
customtkinter
pillow
pywebview